Hopefully fixed pointer arithmetic problems in GCC compilation in cProtocol132
git-svn-id: http://mc-server.googlecode.com/svn/trunk@828 0a769ca7-a7f5-676a-18bf-c427514a06d6

+++ source/Protocol132.cpp
@@ -591,7 +591,7 @@ void cProtocol132::HandleEncryptionKeyResponse(const AString & a_EncKey, const A
 		m_Client->Kick("Hacked client");
 		return;
 	}
-	if (ntohl(*((int *)DecryptedNonce)) != (unsigned)this)
+	if (ntohl(*((int *)DecryptedNonce)) != (unsigned)(uintptr_t)this)
 	{
 		LOGD("Bad nonce value");
 		m_Client->Kick("Hacked client");
